refactor: 형식 통일 및 dto 패키지 변경
