+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,18 +1,25 @@
 package roomescape.controller;
 
 import jakarta.validation.Valid;
+import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import roomescape.dto.AvailableReservationTimeResponse;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.AvailableReservationTimeResponse;
+import roomescape.dto.response.ReservationResponse;
 import roomescape.exception.DuplicateReservationException;
 import roomescape.exception.PastReservationException;
 import roomescape.exception.ReservationTimeConflictException;
 import roomescape.service.ReservationService;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/reservations")
@@ -25,15 +32,8 @@ public class ReservationController {
     }
 
     @GetMapping
-    public ResponseEntity<List<ReservationResponse>> readAllReservations() {
+    public ResponseEntity<List<ReservationResponse>> findAll() {
         return ResponseEntity.ok(reservationService.findAll());
-    }
-
-    @GetMapping("/themes/{themeId}/times")
-    public ResponseEntity<List<AvailableReservationTimeResponse>> readAvailableReservationTimes(
-        @PathVariable Long themeId,
-        @RequestParam String date) {
-        return ResponseEntity.ok(reservationService.findAvailableReservationTime(themeId, date));
     }
 
     @PostMapping
@@ -47,10 +47,17 @@ public class ReservationController {
         return ResponseEntity.noContent().build();
     }
 
+    @GetMapping("/themes/{themeId}/times")
+    public ResponseEntity<List<AvailableReservationTimeResponse>> findAvailableReservationTime(
+            @PathVariable Long themeId,
+            @RequestParam String date) {
+        return ResponseEntity.ok(reservationService.findAvailableReservationTime(themeId, date));
+    }
+
     @ExceptionHandler({
-        PastReservationException.class,
-        DuplicateReservationException.class,
-        ReservationTimeConflictException.class})
+            PastReservationException.class,
+            DuplicateReservationException.class,
+            ReservationTimeConflictException.class})
     public ResponseEntity<String> handleCreateReservationException(final RuntimeException e) {
         return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
     }
