--- src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,14 +1,0 @@
-package roomescape.dto;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-import java.time.LocalDate;
-
-public record ReservationRequest(
-    @Size(min = 1, max = 10) String name,
-    @NotNull LocalDate date,
-    @NotNull Long timeId,
-    @NotNull Long themeId
-) {
-}
