--- src/main/java/roomescape/dto/ThemeRequest.java
@@ -1,10 +1,0 @@
-package roomescape.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ThemeRequest(
-    @NotBlank String name,
-    String description,
-    String thumbnail
-) {
-}
