+++ src/main/java/roomescape/dto/request/ReservationRequest.java
@@ -1,0 +1,13 @@
+package roomescape.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import java.time.LocalDate;
+
+public record ReservationRequest(
+        @Size(min = 1, max = 10) String name,
+        @NotNull LocalDate date,
+        @NotNull Long timeId,
+        @NotNull Long themeId
+) {
+}
