+++ src/main/java/roomescape/dto/request/ReservationTimeRequest.java
@@ -1,10 +1,9 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import jakarta.validation.constraints.NotNull;
-
 import java.time.LocalTime;
 
 public record ReservationTimeRequest(
-    @NotNull LocalTime startAt
+        @NotNull LocalTime startAt
 ) {
 }
