+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -1,0 +1,10 @@
+package roomescape.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ThemeRequest(
+        @NotBlank String name,
+        String description,
+        String thumbnail
+) {
+}
