+++ src/main/java/roomescape/dto/response/ThemeResponse.java
@@ -1,0 +1,20 @@
+package roomescape.dto.response;
+
+import roomescape.entity.Theme;
+
+public record ThemeResponse(
+        Long id,
+        String name,
+        String description,
+        String thumbnail
+) {
+
+    public static ThemeResponse of(Theme theme) {
+        return new ThemeResponse(
+                theme.getId(),
+                theme.getName(),
+                theme.getDescription(),
+                theme.getThumbnail()
+        );
+    }
+}
