+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,16 +1,15 @@
 package roomescape.service;
 
+import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.request.ReservationTimeRequest;
+import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.entity.ReservationTime;
 import roomescape.exception.EntityNotFoundException;
 import roomescape.exception.ReservationExistException;
 import roomescape.exception.ReservationTimeExistException;
 import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
-
-import java.util.List;
 
 @Service
 public class ReservationTimeService {
@@ -23,6 +22,13 @@ public class ReservationTimeService {
         this.reservationTimeDao = reservationTimeDao;
     }
 
+    public List<ReservationTimeResponse> findAll() {
+        List<ReservationTime> times = reservationTimeDao.findAll();
+        return times.stream()
+                .map(ReservationTimeResponse::of)
+                .toList();
+    }
+
     public ReservationTimeResponse add(ReservationTimeRequest requestDto) {
         if (reservationTimeDao.isExistByTime(requestDto.startAt())) {
             throw new ReservationTimeExistException("동일한 시간이 이미 존재합니다.");
@@ -32,21 +38,12 @@ public class ReservationTimeService {
         return ReservationTimeResponse.of(savedReservationTime);
     }
 
-    public List<ReservationTimeResponse> findAll() {
-        List<ReservationTime> times = reservationTimeDao.findAll();
-        return times.stream()
-            .map(ReservationTimeResponse::of)
-            .toList();
-    }
-
     public void deleteById(Long id) {
         if (reservationDao.isExistByTimeId(id)) {
             throw new ReservationExistException("이 시간의 예약이 존재합니다.");
         }
-
         reservationTimeDao.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("삭제할 예약시간이 없습니다."));
-
+                .orElseThrow(() -> new EntityNotFoundException("삭제할 예약시간이 없습니다."));
         reservationTimeDao.deleteById(id);
     }
 }
