+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -1,19 +1,18 @@
 package roomescape.service;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
-import roomescape.entity.Theme;
-import roomescape.repository.ReservationDao;
-import roomescape.repository.ThemeDao;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.List;
 import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import roomescape.dto.request.ThemeRequest;
+import roomescape.dto.response.ThemeResponse;
+import roomescape.entity.Theme;
+import roomescape.repository.ReservationDao;
+import roomescape.repository.ThemeDao;
 
 class ThemeServiceTest {
 
@@ -31,17 +30,17 @@ class ThemeServiceTest {
     @Test
     void 테마를_모두_조회한다() {
         Mockito.when(themeDao.findAll())
-            .thenReturn(List.of(
-                    new Theme(1L, "아이언맨", "", ""),
-                    new Theme(2L, "캡틴아메리카", "", "")
-                )
-            );
+                .thenReturn(List.of(
+                                new Theme(1L, "아이언맨", "", ""),
+                                new Theme(2L, "캡틴아메리카", "", "")
+                        )
+                );
 
         assertThat(themeService.findAll()).isEqualTo(
-            List.of(
-                new ThemeResponse(1L, "아이언맨", "", ""),
-                new ThemeResponse(2L, "캡틴아메리카", "", "")
-            )
+                List.of(
+                        new ThemeResponse(1L, "아이언맨", "", ""),
+                        new ThemeResponse(2L, "캡틴아메리카", "", "")
+                )
         );
     }
 
@@ -50,13 +49,13 @@ class ThemeServiceTest {
         ThemeRequest request = new ThemeRequest("아이언맨", "", "");
 
         Mockito.when(themeDao.isExistByName(Mockito.any()))
-            .thenReturn(false);
+                .thenReturn(false);
 
         Mockito.when(themeDao.save(Mockito.any()))
-            .thenReturn(new Theme(1L, "아이언맨", "", ""));
+                .thenReturn(new Theme(1L, "아이언맨", "", ""));
 
         assertThat(themeService.add(request)).isEqualTo(
-            new ThemeResponse(1L, "아이언맨", "", "")
+                new ThemeResponse(1L, "아이언맨", "", "")
         );
     }
 
@@ -65,10 +64,10 @@ class ThemeServiceTest {
         Long id = 1L;
 
         Mockito.when(reservationDao.isExistByTimeId(Mockito.any()))
-            .thenReturn(false);
+                .thenReturn(false);
 
         Mockito.when(themeDao.findById(Mockito.any()))
-            .thenReturn(Optional.of(new Theme(1L, "", "", "")));
+                .thenReturn(Optional.of(new Theme(1L, "", "", "")));
 
         assertThatCode(() -> themeService.deleteById(id)).doesNotThrowAnyException();
     }
@@ -77,18 +76,18 @@ class ThemeServiceTest {
     @Test
     void 많이_예약된_순서로_테마를_조회한다() {
         Mockito.when(themeDao.sortByRank())
-            .thenReturn(
-                List.of(
-                    new Theme(1L, "아이언맨", "", ""),
-                    new Theme(2L, "캡틴아메리카", "", "")
-                )
-            );
+                .thenReturn(
+                        List.of(
+                                new Theme(1L, "아이언맨", "", ""),
+                                new Theme(2L, "캡틴아메리카", "", "")
+                        )
+                );
 
         assertThat(themeService.sortByRank()).isEqualTo(
-            List.of(
-                new ThemeResponse(1L, "아이언맨", "", ""),
-                new ThemeResponse(2L, "캡틴아메리카", "", "")
-            )
+                List.of(
+                        new ThemeResponse(1L, "아이언맨", "", ""),
+                        new ThemeResponse(2L, "캡틴아메리카", "", "")
+                )
         );
     }
 }
