+++ src/test/resources/application-test.properties
@@ -1,5 +1,4 @@
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 spring.datasource.url=jdbc:h2:mem:database
-# data.sql? ?? ???? ??? ??
 spring.sql.init.mode=never
